+++ Plugin.Notifications.Android/AcrSqliteConnection.cs
@@ -5,14 +5,22 @@ using SQLite;
 
 namespace Plugin.Notifications
 {
-    public class AcrSqliteConnection : SQLiteConnection
+    public class AcrSqliteConnection : SQLiteConnectionWithLock
     {
-        public AcrSqliteConnection() : base(Path.Combine(Environment.Documents.GetPersonalPath(), "notifications.db"), true)
+        public AcrSqliteConnection() :
+            base(new SQLiteConnectionString(
+                Path.Combine(Environment.GetFolderPath(Environment.SpecialFolder.Personal), "notifications.db"),
+                true
+            ), SQLiteOpenFlags.Create | SQLiteOpenFlags.FullMutex | SQLiteOpenFlags.ReadWrite)
         {
             this.CreateTable<DbNotification>();
+            this.CreateTable<DbNotificationMetadata>();
+            this.CreateTable<DbSettings>();
         }
 
 
         public TableQuery<DbNotification> Notifications => this.Table<DbNotification>();
+        public TableQuery<DbNotificationMetadata> NotificationMetadata => this.Table<DbNotificationMetadata>();
+        public TableQuery<DbSettings> Settings => this.Table<DbSettings>();
     }
 }