+++ Plugin.Notifications.Android/AlarmBroadcastReceiver.cs
@@ -9,7 +9,7 @@ namespace Plugin.Notifications
     {
         public override void OnReceive(Context context, Intent intent)
         {
-            //var notificationId = intent.NotificationId();
+            var notificationId = intent.NotificationId();
             //NotificationSettings.Instance.RemoveScheduledId(notificationId);
 
             //// resend without schedule so it goes through normal mechanism