+++ Plugin.Notifications.Android/Helpers.cs
@@ -1,7 +1,6 @@
 using System;
 using Android.App;
 using Android.Content;
-using Newtonsoft.Json;
 
 
 namespace Plugin.Notifications
@@ -28,20 +27,20 @@ namespace Plugin.Notifications
         }
 
 
-        public static Notification ToNotification(this Intent intent)
-        {
-            if (!intent.HasExtra(NOTIFICATION_ID))
-                throw new ArgumentException("Invalid intent package");
+        //public static Notification ToNotification(this Intent intent)
+        //{
+        //    if (!intent.HasExtra(NOTIFICATION_ID))
+        //        throw new ArgumentException("Invalid intent package");
 
 
-            //var notification = JsonConvert.DeserializeObject<Notification>(data);
+        //    //var notification = JsonConvert.DeserializeObject<Notification>(data);
 
-            // nullify date so it doesn't just reschedule
-            //notification.Date = null;
-            //notification.When = null;
-            //return notification;
-            return null;
-        }
+        //    // nullify date so it doesn't just reschedule
+        //    //notification.Date = null;
+        //    //notification.When = null;
+        //    //return notification;
+        //    return null;
+        //}
 
 
         public static int NotificationId(this Intent intent) => intent.GetIntExtra(NOTIFICATION_ID, 0);